refactor: Rename to SessionStorageKeySharer
